+++ src/main/java/com/nusan/nusanapi/repository/PagRepRepository.java
@@ -3,5 +3,5 @@ package com.nusan.nusanapi.repository;
 import com.nusan.nusanapi.model.Report;
 import org.springframework.data.repository.PagingAndSortingRepository;
 
-public interface PagRepRepository extends PagingAndSortingRepository<Report, Integer> {
+public interface PagRepRepository extends PagingAndSortingRepository<Report, Long> {
 }
